backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.3.0
click==8.1.8
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.103.1
greenlet==3.2.0
h11==0.14.0
idna==3.10
passlib==1.7.4
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-telegram==0.16.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.20
starlette==0.27.0
Telethon==1.28.5
typing_extensions==4.13.2
uvicorn==0.23.2
